requirements.txt
fastapi
uvicorn
python-dotenv
semantic-kernel
jinja2